+++ app/reducers/schoolDataReducer.js
@@ -1,17 +1,11 @@
 // @flow
 import {HANDLE_SCHOOL_DATA_DISPLAY } from '../actions/actionTypes';
 import type { Action } from './types';
+import { getSchoolData } from '../database/schoolDB';
 
-const initialLoadState = {
-  'title':'Test Title',
-  'street':'Test Street',
-  'state':'Test State',
-  'country':'Test Country',
-  'year':'Test Year'
-}
+const initialLoadState = {}
 
 const handleDataSchoolReducer = (state = initialLoadState, action) => {
-
   switch (action.type) {
     case HANDLE_SCHOOL_DATA_DISPLAY:
         const { title, street, state, country, year } = action.payload.schoolData[0];
